+++ src/lib.rs
@@ -30,6 +30,7 @@ pub mod widgets;
 
 pub use apply::Apply;
 pub use color::{ColorExt, rgb, rgba};
+pub use style::shared_style::Shadow;
 pub use containers::area::StyledArea;
 pub use containers::column::StyledColumn;
 pub use containers::frame::StyledFrame;