+++ src/prelude.rs
@@ -13,6 +13,7 @@ pub use crate::StyledTheme;
 pub use crate::ThemeExt;
 pub use crate::WebPalette;
 pub use crate::{rgb, rgba};
+pub use crate::Shadow;
 
 pub use crate::StyledArea;
 pub use crate::StyledButton;